refactor(api): move getParticipantsByOrganizationId into prescription bounded context
