+++ api/src/prescription/learner-list/infrastructure/repositories/organization-participant-repository.js
@@ -1,9 +1,9 @@
 import { OrganizationParticipant } from '../../domain/read-models/OrganizationParticipant.js';
-import { knex } from '../../../db/knex-database-connection.js';
-import { fetchPage } from '../utils/knex-utils.js';
-import { filterByFullName } from '../utils/filter-utils.js';
-import { CampaignTypes } from '../../domain/models/CampaignTypes.js';
-import { CampaignParticipationStatuses } from '../../domain/models/CampaignParticipationStatuses.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { fetchPage } from '../../../../../lib/infrastructure/utils/knex-utils.js';
+import { filterByFullName } from '../../../../../lib/infrastructure/utils/filter-utils.js';
+import { CampaignTypes } from '../../../../../lib/domain/models/CampaignTypes.js';
+import { CampaignParticipationStatuses } from '../../../../../lib/domain/models/CampaignParticipationStatuses.js';
 
 async function getParticipantsByOrganizationId({ organizationId, page, filters = {}, sort = {} }) {
   const { count } = await knex
