+++ api/tests/integration/application/organizations/organization-controller_test.js
@@ -20,7 +20,6 @@ describe('Integration | Application | Organizations | organization-controller', 
     sandbox.stub(usecases, 'findCertificationAttestationsForDivision');
     sandbox.stub(usecases, 'findDivisionsByOrganization');
     sandbox.stub(usecases, 'findGroupsByOrganization');
-    sandbox.stub(usecases, 'getPaginatedParticipantsForAnOrganization');
     sandbox.stub(usecases, 'findOrganizationPlacesLot');
 
     sandbox.stub(certificationAttestationPdf, 'getCertificationAttestationsPdfBuffer');
@@ -328,41 +327,6 @@ describe('Integration | Application | Organizations | organization-controller', 
         const response = await httpTestServer.request('GET', `/api/organizations/ABC/groups`);
 
         expect(response.statusCode).to.equal(400);
-      });
-    });
-  });
-
-  describe('#getPaginatedParticipantsForAnOrganization', function () {
-    context('when the organization has participants', function () {
-      it('returns organization participants', async function () {
-        const organizationId = 5678;
-        usecases.getPaginatedParticipantsForAnOrganization
-          .withArgs({ organizationId, page: {}, filters: {}, sort: {} })
-          .resolves({
-            organizationParticipants: [
-              {
-                id: 5678,
-                firstName: 'Mei',
-                lastName: 'Lee',
-              },
-            ],
-            pagination: 1,
-          });
-        securityPreHandlers.checkUserBelongsToOrganization.returns(() => true);
-
-        const response = await httpTestServer.request('GET', `/api/organizations/${organizationId}/participants`);
-
-        expect(response.statusCode).to.equal(200);
-        expect(response.result.data).to.deep.equal([
-          {
-            id: '5678',
-            type: 'organization-participants',
-            attributes: {
-              'first-name': 'Mei',
-              'last-name': 'Lee',
-            },
-          },
-        ]);
       });
     });
   });
