+++ api/tests/prescription/learner-list/unit/domain/usecases/get-paginated-participants-for-an-organization_test.js
@@ -1,7 +1,7 @@
-import { expect, sinon } from '../../../test-helper.js';
-import { getPaginatedParticipantsForAnOrganization } from '../../../../lib/domain/usecases/get-paginated-participants-for-an-organization.js';
+import { sinon, expect } from '../../../../../test-helper.js';
+import { getPaginatedParticipantsForAnOrganization } from '../../../../../../src/prescription/learner-list/domain/usescases/get-paginated-participants-for-an-organization.js';
 
-describe('Unit | UseCase | get-participants-by-organization-id', function () {
+describe('Unit | UseCases | get-paginated-participants-for-an-organization', function () {
   it('should call organizationParticipantRepository', async function () {
     // given
     const organizationId = 90000;
